core: Remove the unit module
